HBASE-22743 : ClientUtils for Demo Client classes (#413)
