+++ hbase-examples/src/main/java/org/apache/hadoop/hbase/rest/RESTDemoClient.java
@@ -19,12 +19,8 @@
 package org.apache.hadoop.hbase.rest;
 
 import java.security.PrivilegedExceptionAction;
-import java.util.HashMap;
-import java.util.Map;
 
 import javax.security.auth.Subject;
-import javax.security.auth.login.AppConfigurationEntry;
-import javax.security.auth.login.Configuration;
 import javax.security.auth.login.LoginContext;
 
 import org.apache.hadoop.hbase.Cell;
@@ -37,6 +33,7 @@ import org.apache.hadoop.hbase.rest.client.Client;
 import org.apache.hadoop.hbase.rest.client.Cluster;
 import org.apache.hadoop.hbase.rest.client.RemoteHTable;
 import org.apache.hadoop.hbase.util.Bytes;
+import org.apache.hadoop.hbase.util.ClientUtils;
 import org.apache.yetus.audience.InterfaceAudience;
 import org.apache.hbase.thirdparty.com.google.common.base.Preconditions;
 
@@ -112,32 +109,7 @@ public class RESTDemoClient {
       return new Subject();
     }
 
-    /*
-     * To authenticate the demo client, kinit should be invoked ahead. Here we try to get the
-     * Kerberos credential from the ticket cache.
-     */
-    LoginContext context = new LoginContext("", new Subject(), null, new Configuration() {
-      @Override
-      public AppConfigurationEntry[] getAppConfigurationEntry(String name) {
-        Map<String, String> options = new HashMap<>();
-        options.put("useKeyTab", "false");
-        options.put("storeKey", "false");
-        options.put("doNotPrompt", "true");
-        options.put("useTicketCache", "true");
-        options.put("renewTGT", "true");
-        options.put("refreshKrb5Config", "true");
-        options.put("isInitiator", "true");
-        String ticketCache = System.getenv("KRB5CCNAME");
-        if (ticketCache != null) {
-          options.put("ticketCache", ticketCache);
-        }
-        options.put("debug", "true");
-
-        return new AppConfigurationEntry[] {
-          new AppConfigurationEntry("com.sun.security.auth.module.Krb5LoginModule",
-                AppConfigurationEntry.LoginModuleControlFlag.REQUIRED, options) };
-      }
-    });
+    LoginContext context = ClientUtils.getLoginContext();
     context.login();
     return context.getSubject();
   }
